respond to PR feedback
