+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.2] - 2025-03-03
+## [1.0.2] - 2025-03-04
 
 - Pass metadata in headers for query requests.
