+++ src/cleanlab_codex/internal/analytics.py
@@ -1,15 +1,22 @@
 from __future__ import annotations
 
+from enum import Enum
+
 from cleanlab_codex.__about__ import __version__ as package_version
 
 
-class AnalyticsMetadata:
-    def __init__(
-        self, *, integration_type: str, package_version: str = package_version, source: str = "cleanlab-codex-python"
-    ):
+class IntegrationType(str, Enum):
+    """Types of integrations that can use the SDK."""
+
+    BACKUP = "backup"
+    TOOL = "tool"
+
+
+class _AnalyticsMetadata:
+    def __init__(self, *, integration_type: IntegrationType):
         self._integration_type = integration_type
         self._package_version = package_version
-        self._source = source
+        self._source = "cleanlab-codex-python"
 
     def to_headers(self) -> dict[str, str]:
         return {
